python-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.7.4
websockets==12.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
